prow: Fix branch protection for CAPM3
This fixes branch protection settings for cluster-api-provider-metal3.
The previous configuration made master require both "test-integration"
and "test-v1a4-integration".  The requirements are the union of parent
and child lists, instead of only the child list.

Instead, we now specify the required contexts for this repo on a per
branch basis.

Reported-by: maelk <mael.kimmerlin@est.tech>