Get Ironic Python Agent from Artifactory
In order to reduce dependency on upstream services (e.g. in the event of
network degradation between Sweden and US West, where the default Ironic
Python Agent tarball is hosted), set IPA_BAREURI to download IPA from
Nordix Artifactory in CI. This has the added benefit of potentially
reducing build times, because we'll be downloading from the same
infrastructure on which the CI is hosted.