Fix jobs after change to Fra1 region
The previous pattern matching logic for UPDATED_BRANCH being main or
master doesn't work in bash (only in zsh). This is updated so that the
OpenStack environment variables are set correctly now if we're running
the job on a main branch.

Furthermore, logging is added so that the region is echoed in all runs.

Additionally, the cleanup job was broken by the previous change. It is
now updated to run resource cleanup in both regions.