Pick a random floating IP for Frankfurt
If two jobs run at the same time, they will take the same IP address at
the top of the floating IP list. By taking a random IP address, we
reduce the chances of an IP address collision in CI.

Ideally, we create and destroy floating IPs as part of the job in order
to avoid this race condition, but this requires a bit more work, I've
added a TODO for this.