Prow: add cherrypicker external plugin
This adds cherry-pick external plugin.
Short description from https://prow.k8s.io/plugins

"The cherrypick plugin is used for cherrypicking
PRs across branches. For every successful
cherrypick invocation a new PR is opened
against the target branch and assigned to
the requestor. If the parent PR contains
a release note, it is copied to the cherrypick
PR."