Add linter CI for YAML files
Currently we don't lint YAML files. As a result, sometimes we see
issues with some YAML code being merged and the issue is identified
later on. Example, trailing spaces which might breake the
configuration file of Prow.

This patch adds yamllinter based GitHub action so that we verify
correctness of YAML files on every pull request submission, before
landing on the main branch. As a start, trailing space, key-duplicates
and truthy rules are enabled, and the list can be extended based on
the need.