New Prow instance and docs
The Metal3 Prow instance has been replaced with a new that runs in a
Cluster API cluster on OpenStack (Cleura). It is using the S3 API in
Cleura for object storage (instead of MinIO in front of Azure for the
old instance).

This commit also includes basic docs for how to set up the cluster and
deploy Prow.

Main changes:

- Run Prow in a CAPI/CAPO cluster instead of hand crafted kubeadm
  cluster
- Cinder volume for ghproxy instead of host path on the node
- HTTP01 challenge for certificates instead of DNS01 since we no longer
  have direct API access for DNS management.
- Removed old ingress subdomain that was unused
  (hook.apps.test.metal3.io).
- S3 storage in Cleura instead of MinIO in front of Azure
- New kustomizations for Prow, "infra" and the cluster itself.
- Moved label sync CronJob to the prow namespace (previously default).
- Changes in folder/file structure. Some files have been moved without
  changes.