Prow: Upgrade to k8s v1.26.6
This upgrades the Kubernetes cluster to v1.26.6 and includes a "safety
handle" that was recently added in CAPO templates, namely to set the
provider ID directly through the kubelet. This prevents potential issues
from differences between the OpenStackMachine and server names. If they
differ for some reason, then the cloud controller will not be able to
set the correct provider ID. By telling the kubelet to set the provider
ID directly, we bypass this completely.