build capm3 if top-level go.mod or main.go changes
Currently this job only builds if api/ or test/ or Makefile change,
but "make build" also builds the main binary from top-level. It might
be confusing to users to see this job pass/fail differently from other
jobs that build top-level as well, so triggering the build job here
as well.