Trigger BMO e2e tests through prow
This is to verify that the jenkins operator for prow works as expected.
The idea is to trigger jenkins jobs through prow instead of through the
ghprb jenkins plugin. When we have verified that this works, we can
continue with the rest of the jenkins jobs and remove the ghprb
triggers. Then we can also update the required status checks and make
the prow-jenkins jobs required.

Signed-off-by: Lennart Jern <lennart.jern@est.tech>