Make BMO e2e prow job required
This removes the required status context for BMO e2e tests and instead
makes the prow job required. It also adds another optional job.

It also removes the report template because this is already covered by
the report template set for plank. This config is currently causing
double messages when a job fails (one for jenkins and one for plank).

Signed-off-by: Lennart Jern <lennart.jern@est.tech>