Add cluster-autoscaler for prow cluster
This adds the cluster-autoscaler to the prow cluster and configures the
MachineDeployment to be managed by it. It is configured to scale between
1 and 5 replicas.
The deployment manifest is based on the upstream example. The autoscaler
is configured to run on control-plane nodes only, as recommended by
upstream. It has the necessary RBAC to manage resources in the prow
cluster itself but nothing else.

Signed-off-by: Lennart Jern <lennart.jern@est.tech>