Prow: Enable blunderbuss
Blunderbuss is a Prow plugin that selects and asks reviewers to review
PRs. Unfortunately the docs are a bit on the light side for how to
actually configure it. The available docs can be found here:
https://docs.prow.k8s.io/docs/components/plugins/approve/approvers/#blunderbuss-and-reviewers

I have used the test-infra configuration as example to learn how to
actually configure it.
https://github.com/kubernetes/test-infra/blob/75d0608061cac39e624fad37b4f082dfa84e7ba5/config/prow/plugins.yaml#L331-L333

The configuration in this commit should select 2 reviewers for each PR
for all metal3.io PRs and respect the GitHub status. (If you set
yourself as busy it won't pick you.)

Signed-off-by: Lennart Jern <lennart.jern@est.tech>