Fix git credentials ID
We have been using the wrong credentials ID, resulting in "not found"
errors (but still working without credentials). The thing is that the
git plugin only supports username/password, NOT token. We have been
trying to use a token for checking out the tested repo.
We have been using the correct credentials for checking out
project-infra though.

This PR changes all pipelines to use the same credentials ID we always
used for checking out project-infra.

Signed-off-by: Lennart Jern <lennart.jern@est.tech>