Prow: Skip job report from jenkins-operator
We currently see double comments on github when jenkins jobs fail. I
think the issue is that the jenkins-operator reports the failure once
and prow itself reports it a second time (since the job also has a
corresponding prowjob). If this is accurate, we can solve the issue by
disabling reports from the jenkins-operator, which is what this commit
does.

Signed-off-by: Lennart Jern <lennart.jern@est.tech>