Bump golang to v1.22 for CAPM3, BMO and IPAM prow jobs
This PR bumps main branch golang version to go v1.22 for CAPM3, BMO and IPAM prow jobs. Only remaining repo in this config file is metal3-io/ironic-standalone-operator whic will still be using go 1.21 in main branch prow tests. We need to tackle that in a separate PR once a consensus is reached for that repo.
Signed-off-by: Kashif Khan <kashif.khan@est.tech>