Allow uploading the latest image under a common name
This PR makes sure that the latest image created in Openstack is pushed under a common name. This will make sure that the image being used in Jenkins workers remain active all the time. When replacing an old image with the latest, this commit also renames the old image based on its original name which is build date prefix and this will be added in the images as an image_name property.

Signed-off-by: Kashif Khan <kashif.khan@est.tech>