update basic-checks baseimage when building it
Update baseimage when we build the basic-checks image. While we usually
bump to quite new version, we also have repository_dispatch trigger in
the workflow, so we can rebuild the image in which case baseimage is
old and should get updated.

Quay is reporting a lot of vulnerabilities for this image already.

Signed-off-by: Tuomo Tanskanen <tuomo.tanskanen@est.tech>