Prow: Use GCR for commenter and label_sync images
These components are not maintained in the new prow repo and have
apparently not been pushed to the same staging registry. Instead they
are pushed to the test-infra staging registry in GCR.

This commit changes the registry for the commenter and label_sync jobs
and updates the autobumper to detect these images as well. I also had to
bump the tag of the images since they have not pushed old images to the
same registry.

Signed-off-by: Lennart Jern <lennart.jern@est.tech>