remove dracut from the IPA rootfs
this commit:
  - removes the dracut packages from the rootfs

The size of the IPA image is slowly but steadily growing, and in order to
minimize the size of the image some packages could be removed.

Although dracut is used at some point in the DIB build process there  is no
point leaving it in the root file system after the IPA build process is
finished. IPA user space does not rely on dracut, systemd can handle everything
on it's own.

Signed-off-by: Adam Rozman <adam.rozman@est.tech>