group github actions
Change dependabot config not to try major bumps of github actions.
We now have grouped bumps to lessen the noise from dependabot, and
if in that group, there is a major bump (like right now a
golangci-lint v2 major bump in some othe repos) which isn't compatible
as is and requires manual changes, it blocks bumping all the other
actions as well.

Signed-off-by: Tuomo Tanskanen <tuomo.tanskanen@est.tech>