Fix image build refspec and upload
When triggering jobs on PRs, we need to include the PR commit(s) in the
refspec. Otherwise we cannot checkout the relevant code. Pre-build-merge
is also removed to keep the behavior similar to other jobs.
This commit also avoids uploading images from PR tests and making them "latest".
It also fixes the labels for the CI image verification step so that the
label is separate for ubuntu and centos images.

Last but not least, this tweaks the logic for image retention.
CI images are uploaded as metal3-ci-<os>-staging for verification.
In periodic jobs, the staging images are then renamed to *-latest after
successful verification. This is not done in PR jobs.

Any existing *-staging images are deleted before uploading a new.
Otherwise we risk mixing them up with the previously used CI images and
fill the retention queue with staging images instead of the real
"production" images because the logic of the upload script is to rename
them back to their "original" name when there are conflicts.

Signed-off-by: Lennart Jern <lennart.jern@est.tech>