add image SBOM support for container-image-build
Add image SBOM support for container-image-build.yml.

At the start, it will be disabled by default, and has to be enabled
by the matching build-images-action.yml, image by image. Once it is
proven to be working, we can make it default.

We'll use "bom" tool from kubernetes-sigs to generate the SPDX, and
cosign to attest and upload it to the registry.

Signed-off-by: Tuomo Tanskanen <tuomo.tanskanen@est.tech>