extend prow_test_dumper.py to check whole branch config
We now just check the test suites are correct in release branch
protections. This PR extens the checks to verify the branch protections
exist, and branches that are not supported are locked. We also scream
if locked branches have prow tests left, so we can clean them up, and
we scream if we find non-locked branch that is missing them.

We also check the token provided has right scope to check these, and
do a test to see if we have role to read branch protections.

Signed-off-by: Tuomo Tanskanen <tuomo.tanskanen@est.tech>